tests: move landscape pending computer cleanup to after_all step
